requirements.txt
langchain-community
langchain-openai
openai
SQLAlchemy
streamlit

